requirements.txt
Package               Version
--------------------- -------
asgiref               3.8.1
Django                5.0.7
django-webpack-loader 3.1.0
djangorestframework   3.15.2
pillow                10.4.0
pip                   24.1.2
sqlparse              0.5.1
tzdata                2024.1
